Add files via upload
Essentially v1.2

Added persistence for gain mode, persistence for more than one setting didn't work. Feel free to make a pull request to fix that

Changed different gain modes to be more musical and have more variation. Can be used as standalone distortion now in addition to reverb

Implemented ripped speaker distortion, Bogaudio LMTR, foldback distortion, and more

If in gain control mode not confirming will lock mod depth to 3.125% and the mod depth knob will now control mod shape

Removed some premature optimizations in the interpolating delay and in the low and high pass filters that were making it not sound perfectly faithful to the original

Double precision instead of single precision, this was another premature optimization, the STM32H750 can handle it with the double precision FPU core

Output gain control now controls the distortion settings instead, clip threshold for example

+++ dsp/delays/InterpDelay.cpp
@@ -3,4 +3,4 @@
 float DSY_SDRAM_BSS sdramData[50][144000];
 unsigned int count = 0;
 bool triggerClear;
-float hold = 0.f;
+double hold = 0.;
