+++ dsp/modulation/LFO.hpp
@@ -10,20 +10,20 @@
 
 class TriSawLFO {
 public:
-    TriSawLFO(float sampleRate = 32000.0f, float frequency = 1.0f) {
-        phase = 0.0f;
-        _output = 0.0f;
+    TriSawLFO(double sampleRate = 32000.0, double frequency = 1.0) {
+        phase = 0.0;
+        _output = 0.0;
         _sampleRate = sampleRate;
         _1_sampleRate = 1 / sampleRate;
-        _step = 0.0f;
+        _step = 0.0;
         _rising = true;
         setFrequency(frequency);
-        setRevPoint(0.5f);
+        setRevPoint(0.5);
     }
 
-    inline float process() {
-        if(_step > 1.0f) {
-            _step -= 1.0f;
+    inline double process() {
+        if(_step > 1.0) {
+            _step -= 1.0;
             _rising = true;
         }
 
@@ -39,12 +39,12 @@ public:
         }
 
         _step += _stepSize;
-        _output *= 2.0f;
-        _output -= 1.0f;
+        _output *= 2.0;
+        _output -= 1.0;
         return _output;
     }
 
-    inline void setFrequency(const float &frequency) {
+    inline void setFrequency(const double &frequency) {
         if (frequency == _frequency) {
             return;
         }
@@ -52,41 +52,41 @@ public:
         calcStepSize();
     }
 
-    inline void setRevPoint(const float &revPoint) {
+    inline void setRevPoint(const double &revPoint) {
         _revPoint = revPoint;
-        if(_revPoint < 0.0001f) {
-            _revPoint = 0.0001f;
+        if(_revPoint < 0.0001) {
+            _revPoint = 0.0001;
         }
-        if(_revPoint > 0.999f) {
-            _revPoint = 0.999f;
+        if(_revPoint > 0.999) {
+            _revPoint = 0.999;
         }
 
-        _riseRate = 1.0f / _revPoint;
-        _fallRate = -1.0f / (1.0f - _revPoint);
+        _riseRate = 1.0 / _revPoint;
+        _fallRate = -1.0 / (1.0 - _revPoint);
     }
 
-    void setSamplerate(float sampleRate) {
+    void setSamplerate(double sampleRate) {
         _sampleRate = sampleRate;
         _1_sampleRate = 1 / sampleRate;
         calcStepSize();
     }
 
-    inline float getOutput() const {
+    inline double getOutput() const {
         return _output;
     }
 
-    float phase;
+    double phase;
 
 private:
-    float _output;
-    float _sampleRate;
-    float _1_sampleRate;
-    float _frequency = 0.0f;
-    float _revPoint;
-    float _riseRate;
-    float _fallRate;
-    float _step;
-    float _stepSize;
+    double _output;
+    double _sampleRate;
+    double _1_sampleRate;
+    double _frequency = 0.0;
+    double _revPoint;
+    double _riseRate;
+    double _fallRate;
+    double _step;
+    double _stepSize;
     bool _rising;
 
     inline void calcStepSize() {
